rename DesignSystem - take out Fonts apply DRY on NavigationControllers
