+++ App/CourseApp/SharedUI/TextType.swift
@@ -1,5 +1,5 @@
 //
-//  DesignSystem.swift
+//  TextType.swift
 //  CourseApp
 //
 //  Created by Marcel Mravec on 17.05.2024.
@@ -19,6 +19,15 @@ enum FontType: String {
     case regular = "Poppins-Regular"
     case bold = "Poppins-Bold"
     case mediumItalic = "Poppins-MediumItalic"
+}
+
+struct TextTypeModifier: ViewModifier {
+    let textType: TextType
+    func body(content: Content) -> some View {
+        content
+            .font(textType.font)
+            .foregroundColor(textType.color)
+    }
 }
 
 enum TextType {
@@ -41,32 +50,6 @@ enum TextType {
         default:
             .gray
         }
-    }
-}
-
-struct TextTypeModifier: ViewModifier {
-    let textType: TextType
-    func body(content: Content) -> some View {
-        content
-            .font(textType.font)
-            .foregroundColor(textType.color)
-    }
-}
-
-extension UIFont {
-    static func regular(with size: FontSize) -> UIFont {
-        UIFont(name: FontType.regular.rawValue, size: size.rawValue)!
-    }
-    static func bold(with size: FontSize) -> UIFont {
-        UIFont(name: FontType.bold.rawValue, size: size.rawValue)!
-    }
-}
-extension Font {
-    static func regular(with size: FontSize) -> Font {
-        Font.custom(FontType.regular.rawValue, size: size.rawValue)
-    }
-    static func bold(with size: FontSize) -> Font {
-        Font.custom(FontType.bold.rawValue, size: size.rawValue)
     }
 }
