+++ App/CourseApp/Views/SwiftUI/HomeViewSwiftUI.swift
@@ -5,6 +5,7 @@
 //  Created by Marcel Mravec on 16.05.2024.
 //
 
+// TODO: Not used atm - remove?
 import os
 import SwiftUI
